reintegrados os Changesets: 147562, 147606 e 147609 ref. a Branch(Bugs) para a Main. Imported from Changeset 147668

+++ Ponto/BLL/CalculoMarcacoes/TotalizadorHoras.cs
@@ -17,7 +17,7 @@ namespace BLL
         public int QuantidadeExtraMin { get; set; }
         public short TipoAcumulo { get; set; }
         public int PercentualExtraSegundo { get; set; }
-        public decimal? PercentualExtraNoturna { get; set; }
+        public int? PercentualExtraNoturna { get; set; }
         public string QuantidadeExtraNoturna { get; set; }
         public int QuantidadeExtraNoturnaMin { get; set; }
         public int? PercentualExtraSegundoNoturna { get; set; }
